+++ VMDeploy.Management/changelog.md
@@ -2,6 +2,11 @@
 
 ## ???
 
++ Major: Redesign to support multiple SCVMM Servers
++ New: Command Register-VMManSCVMM
++ New: Command Get-VMManSCVMM
++ New: Command Unregister-VMManSCVMM
++ Upd: Configuration: Template - added property "VmmServer"
 + Upd: Configuration: Template - added property "DynamicHardwareProfile"
 + Upd: Configuration: Template - added property "DynamicGuestOSProfile"
 + Upd: JEA Endpoint - added new commands from VMDeploy.Orchestrator